feat(objectql): materialize seed rows on EVERY publish path (seedApplied)
Publishing a `seed` draft only promoted its metadata — whether rows actually
landed depended on WHICH route you published through. The package route
(/packages/:id/publish-drafts) applied seeds at the dispatcher level and
swallowed failures silently; the per-ref publish (/meta/:type/:name/publish —
what the home banner uses) never applied them at all. Result: "Published!"
with an app full of empty tables.

Move SeedLoaderService into @objectstack/objectql (runtime keeps a re-export
shim) and apply seeds where every path converges — the protocol itself:

- publishMetaItem: publishing a single `seed` materializes its rows from the
  just-promoted body (no read-back, no org-scope resolution pitfalls) and
  reports under `seedApplied`. Best-effort: a seed problem NEVER fails the
  publish, but it is now LOUD in the response instead of invisible.
- publishPackageDrafts: structure publishes first, seeds LAST (their objects'
  tables must exist), with every seed body batch-applied in ONE loader pass so
  cross-seed references resolve regardless of draft order. Per-item self-apply
  is suppressed (_skipSeedApply) so nothing double-inserts.
- http-dispatcher publish-drafts route: falls back to its route-level apply
  ONLY when the protocol didn't self-apply (custom protocol back-compat).

objectql 546 + runtime 381 tests green; new coverage for ordering, batch
single-pass, _skipSeedApply, non-seed publishes, and a real-loader smoke test.
